+++ Bolt-Core/src/Graphic/GlRenderer.cpp
@@ -17,33 +17,33 @@ namespace bolt {
 
     ShaderProgram* Renderer::getDefaultShader() const { return this->m_shader.get(); }
 
-	void Renderer::drawArrays(const VertexArray &vao, const u32 &mode, const i32 &first, const i32 &count) {
+	void Renderer::drawArrays(const VertexArray &vao, const u32 &mode, const i32 &first, const size_t &count) {
 		vao.bind();
-		glDrawArrays(mode, first, count);
+		glDrawArrays(mode, first, static_cast<i32>(count));
 	}
 
-	void Renderer::drawElements(const VertexArray &vao, const u32 &mode, const i32 &count, const u32 &type, const void *indices) {
+	void Renderer::drawElements(const VertexArray &vao, const u32 &mode, const size_t &count, const u32 &type, const void *indices) {
 		vao.bind();
-		glDrawElements(mode, count, type, indices);
+		glDrawElements(mode, static_cast<i32>(count), type, indices);
 	}
 
-	void Renderer::drawArraysLines(const VertexArray &vao, const i32 &count, const i32 &first) {
+	void Renderer::drawArraysLines(const VertexArray &vao, const size_t &count, const i32 &first) {
 		this->drawArrays(vao, GL_LINES, first, count);
 	}
 
-	void Renderer::drawElementsLines(const VertexArray &vao, const i32 &count) {
+	void Renderer::drawElementsLines(const VertexArray &vao, const size_t &count) {
 		this->drawElements(vao, GL_LINES, count, GL_UNSIGNED_INT, nullptr);
 	}
 
-	void Renderer::drawArraysTriangles(const VertexArray &vao, const i32 &count, const i32 &first) {
+	void Renderer::drawArraysTriangles(const VertexArray &vao, const size_t &count, const i32 &first) {
 		this->drawArrays(vao, GL_TRIANGLES, first, count);
 	}
 
-	void Renderer::drawArraysTriangleFan(const VertexArray &vao, const i32 &count, const i32 &first) {
+	void Renderer::drawArraysTriangleFan(const VertexArray &vao, const size_t &count, const i32 &first) {
 		this->drawArrays(vao, GL_TRIANGLE_FAN, first, count);
 	}
 
-	void Renderer::drawElementsTriangles(const VertexArray &vao, const i32 &count) {
+	void Renderer::drawElementsTriangles(const VertexArray &vao, const size_t &count) {
 		this->drawElements(vao, GL_TRIANGLES, count, GL_UNSIGNED_INT, nullptr);
 	}
 } // namespace bolt
